Added MCU_BASENAME_nnn to progmem_string_data.h, replacing basename.h with a more performant version.
Various of the macros now take or build the name of a macro function to be
evaluated, commonly the name of the macro used to produce a leaf fragment,
either _PSD_STRFRAG_TYPE or _PSD_PATH_FRAG_TYPE. This avoids a lot of
unnecessary duplication.

Renamed the progmem_data namespace to progmem_string_data so that it doesn't
conflict with the namespace in inline_literal.h.

Tidied up comments and tests.