Prefer using ProgmemString or ProgmemStringView over Literal.
Add has_progmem_char_array.h, with a trait of the same name.

Rename print_to_trait.h to has_print_to.h, so that all of the
"non-standard" traits are defined in ~adjacent files, and not
in type_traits.h (i.e. so that we can replace type_traits.h with
the standard <type_traits> where available).