Fix hex escaping support in the case of a hex digit following a hex escaped character, and escape multiple question marks in a row to avoid emitting trigraphs.
Don't emit \t for a tab character, instead use hex escaping. Limit that style
to just \n and \r as I expect that most uses of control character logging to
be for binary data where the user wants to interpret the value of the bytes.

Improve testing of of hex escape.