Add a compile-time option (the default) to use FNV-1a into JitterRandom.
Move the WaitForInterrupt associated with the starting of each Timer/Counter
to just before we start it, rather than just after. This means we don't do
any extra waits, and we do wait before starting the first T/C.

Change the params of JitterRandom::random32 so that the caller can specify
the minimum number of watchdog interrupts to wait for, and the minimum elapsed
time over which to acquire data; both must be satisfied before the function
returns a value.

Modify jitter_random_exercise.ino to reboot after gathering the
specified data, at which point it will start over.