Remove MessageSinkBase, and unneeded ctors for LogSink and CheckSink.
Remove extra announcement from the CheckSink dtor. Instead just assume
that we're always going to exit the program/restart the sketch. This
allows us to remove the data members of MessageSinkBase, which was the
reason it existed, and to remove the data member of CheckSink.

Add a doc with ideas about selective checking and logging.